Major bug fix in bspline_longslit. It was producing garbage results whenever any breakpoints were being rejected. I've now checked that it produces identical results to idl in these cases.

+++ pypit/core/pydl.py
@@ -10,6 +10,7 @@ from warnings import warn
 from pypit import msgs
 from pypit import ardebug as debugger
 
+
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 # -*- coding: utf-8 -*-
 """This module corresponds to the image directory in idlutils.
@@ -416,15 +417,15 @@ class bspline(object):
         n = nbkpt - self.nord
         gb = self.breakpoints[self.mask]
         bw = self.npoly*self.nord
-        lower = np.zeros((n - self.nord + 1,), dtype='i4')
-        upper = np.zeros((n - self.nord + 1,), dtype='i4') - 1
+        lower = np.zeros((n - self.nord + 1,), dtype=int)
+        upper = np.zeros((n - self.nord + 1,), dtype=int) - 1
         indx = self.intrv(x)
         bf1 = self.bsplvn(x, indx)
         action = bf1
-        aa = uniq(indx, np.arange(indx.size, dtype='i4'))
+        aa = uniq(indx, np.arange(indx.size, dtype=int))
         upper[indx[aa]-self.nord+1] = aa
         rindx = indx[::-1]
-        bb = uniq(rindx, np.arange(rindx.size, dtype='i4'))
+        bb = uniq(rindx, np.arange(rindx.size, dtype=int))
         lower[rindx[bb]-self.nord+1] = nx - bb - 1
         if x2 is not None:
             if x2.size != nx:
@@ -471,7 +472,7 @@ class bspline(object):
         """
         gb = self.breakpoints[self.mask]
         n = gb.size - self.nord
-        indx = np.zeros((x.size,), dtype='i4')
+        indx = np.zeros((x.size,), dtype=int)
         ileft = self.nord - 1
         for i in range(x.size):
             while x[i] > gb[ileft+1] and ileft < n - 1:
